requirements.txt
GitPython==3.1.31
python-on-whales==0.60.0
